feat(apollo-vertex): add Components overview page
