+++ apps/apollo-vertex/app/components/overview/page.mdx
@@ -1,0 +1,7 @@
+import { OverviewGrid } from '../_components/overview-grid'
+
+# Components
+
+Browse all available components in the Apollo Vertex design system.
+
+<OverviewGrid />
